Rewrite 2015/24 in ocaml
This solution is not as complete as the python one as it does not ensure
the group with the lowest entanglement forms a valid set of groups,
i.e. remaining packages can be successfully allocated.

This does not appear necessary to matter with this specific input.
Changing the python solution to have this same assumption brings down
its runtime to about a quarter second from 6. Ocaml solution runs in
nearly the same time.