Time each solution in the implementation
# Problem

Our main run script currently gets the runtime of each solution
by timing how long it takes the bash run command for the solution
to complete.

This is close enough for languages like Python which don't have much
pre-processing or general initialization overhead but worse for
languages with a heavy runtime. Plus the general overhead of executing
a subprocess command in Python, though this is not the most significant.

Ideally we could get something closer to the actual amount of time
the solution takes to execute.

# Solution

Update the Answer module for each language with a timer function.

This should be used to wrap the main method of each solution.

This executes the main method and print out the runtime in nanoseconds.

We can then extract this value in our run script and use this more
accurate runtime value instead, this portion is not yet implemented.

Created the function for each language, updated the 2021/01 solutions
to use it to make sure it works. Updated the templates for each language
accordingly.