Update graphs after adding terminals to developer tools
## Details

Turns out a good amount of language overhead and a nice chunk of the
measured runtime came from the fact that MacOS runs `XprotectService` on
all executions of all binaries (which is not a bad idea).

By adding my terminal emulators to the list of applications under
`Developer Tools` this overhead essentially disappears, nice, and has a
measured impact.